- homepage and main page created as jsp - added errorHomePage.jsp - created LoginServlet - fixed some issues to the Account and DebitAccount classes - fixed incorrect closing of the DBManager class - updated DBUserDAO class with adding the user_id - removed / fixed some of the methods in the IAccount interface

+++ src/model/interfaces/IAccount.java
@@ -7,11 +7,15 @@ import java.util.Currency;
 
 public interface IAccount {
 
-    void withdrawMoney(double money);
+    void withdrawMoney(BigDecimal money);
 
-    void insertMoney(double money);
+    void insertMoney(BigDecimal money);
 
-    double getAmount();
+//    void generateTask(User user);
+
+//    Currency getCurrency();
+
+    BigDecimal getAmount();
 
     String getAccountName();
 }
